Installer database adjustments
We can't say we officially support PostgreSQL, unfortunately. There
are too many database calls with MySQL specific syntax. This would be
desirable for a 2.0 release, but too much work while maintaining 1.x.

The main difficulty is that we're using PEAR::DB which is aging. If
that's exchanged, maybe we could use PDO or something.